requirements.txt
Django==4.1.6
djangorestframework==3.14.0
psycopg2==2.9.5
pillow==9.4.0
django-cors-headers==3.10.1